Open google meet app instead of creating a new meeting on Android
