+++ src/components/VideoChatButtonAndMenu/BaseVideoChatButtonAndMenu.js
@@ -3,34 +3,24 @@ import React, {Component} from 'react';
 import {
     View, Pressable, Dimensions, Linking,
 } from 'react-native';
-import PropTypes from 'prop-types';
-import Icon from './Icon';
-import * as Expensicons from './Icon/Expensicons';
-import Popover from './Popover';
-import MenuItem from './MenuItem';
-import ZoomIcon from '../../assets/images/zoom-icon.svg';
-import GoogleMeetIcon from '../../assets/images/google-meet.svg';
-import CONST from '../CONST';
-import styles from '../styles/styles';
-import themeColors from '../styles/themes/default';
-import withWindowDimensions, {windowDimensionsPropTypes} from './withWindowDimensions';
-import withLocalize, {withLocalizePropTypes} from './withLocalize';
-import compose from '../libs/compose';
-import Navigation from '../libs/Navigation/Navigation';
-import ROUTES from '../ROUTES';
-import Tooltip from './Tooltip';
+import Icon from '../Icon';
+import * as Expensicons from '../Icon/Expensicons';
+import Popover from '../Popover';
+import MenuItem from '../MenuItem';
+import ZoomIcon from '../../../assets/images/zoom-icon.svg';
+import GoogleMeetIcon from '../../../assets/images/google-meet.svg';
+import CONST from '../../CONST';
+import styles from '../../styles/styles';
+import themeColors from '../../styles/themes/default';
+import withWindowDimensions from '../withWindowDimensions';
+import withLocalize from '../withLocalize';
+import compose from '../../libs/compose';
+import Navigation from '../../libs/Navigation/Navigation';
+import ROUTES from '../../ROUTES';
+import Tooltip from '../Tooltip';
+import {propTypes, defaultProps} from './videoChatButtonAndMenuPropTypes';
 
-const propTypes = {
-    ...withLocalizePropTypes,
-    ...windowDimensionsPropTypes,
-    isConcierge: PropTypes.bool,
-};
-
-const defaultProps = {
-    isConcierge: false,
-};
-
-class VideoChatButtonAndMenu extends Component {
+class BaseVideoChatButtonAndMenu extends Component {
     constructor(props) {
         super(props);
 
@@ -53,7 +43,7 @@ class VideoChatButtonAndMenu extends Component {
                 text: props.translate('videoChatButtonAndMenu.googleMeet'),
                 onPress: () => {
                     this.toggleVideoChatMenu();
-                    Linking.openURL(CONST.NEW_GOOGLE_MEET_MEETING_URL);
+                    Linking.openURL(this.props.newGoogleMeetingUrl || CONST.NEW_GOOGLE_MEET_MEETING_URL);
                 },
             },
         ];
@@ -148,10 +138,10 @@ class VideoChatButtonAndMenu extends Component {
     }
 }
 
-VideoChatButtonAndMenu.propTypes = propTypes;
-VideoChatButtonAndMenu.defaultProps = defaultProps;
+BaseVideoChatButtonAndMenu.propTypes = propTypes;
+BaseVideoChatButtonAndMenu.defaultProps = defaultProps;
 
 export default compose(
     withWindowDimensions,
     withLocalize,
-)(VideoChatButtonAndMenu);
+)(BaseVideoChatButtonAndMenu);
